requirements.txt
httpx==0.24.1
aiogram>=3.7.0
python-dotenv==1.0.0
loguru==0.7.2 